requirements.txt
torch>=2.0
torchvision>=0.15
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
scipy>=1.11
matplotlib>=3.7
seaborn>=0.12
Pillow>=10.0
opencv-python>=4.8
umap-learn>=0.5
